pyproject.toml
# SPDX-FileCopyrightText: Copyright (C) Nicolas Lamirault <nicolas.lamirault@gmail.com>
# SPDX-License-Identifier: Apache-2.0

[project]
name = "ai-llm-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "langchain>=0.3.27",
  "langchain-anthropic>=0.3.21",
  "langchain-openai>=0.3.21",
  "langsmith[otel]>=0.4.31",
  "opentelemetry-api==1.37.0",
  "opentelemetry-distro==0.58b0",
  "opentelemetry-exporter-otlp==1.37.0",
  "opentelemetry-exporter-otlp-proto-grpc==1.37.0",
  "opentelemetry-exporter-otlp-proto-http==1.37.0",
  "opentelemetry-instrumentation-langchain==0.47.3",
  "opentelemetry-sdk==1.37.0",
  "opentelemetry-semantic-conventions>=0.58b0",
  # "paid-python>=0.0.5",
]
